backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
anthropic==0.43.0
apscheduler==3.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.3
